Integrate Get APIs for Partners and Ledger
